[MINOR] Split ITConversionController so its tests can use the fork pool
Failsafe is configured with reuseForks=false, so a test class occupies
exactly one fork and its tests run serially there no matter how high
forkCount is. ITConversionController held 13 test methods expanding to
44 invocations, and at 1003s it was the single longest thing in the
build - xtable-core takes 18:05 of a 27:22 CI run, and the class alone
accounts for most of that. No amount of extra forks, caching or runner
capacity can shorten it while it is one class.

Move the shared Spark fixture and the assertion helpers into a new
abstract ConversionControllerTestBase and split the tests across five
classes grouped by what they exercise:

  ITConversionController                    testVariousOperations
  ITConversionControllerIncrementalSync     time travel, single-format,
                                            out-of-sync, no-op, retention
  ITConversionControllerSchemaEvolution     UUID columns, Delta column
                                            mapping, snapshot recovery
  ITConversionControllerConcurrentWrites    concurrent inserts, compaction
  ITConversionControllerPartitioning        partition transforms

Test bodies and assertions are unchanged; 13 methods and 44 invocations
before and after. The base class is deliberately named so that it
matches neither failsafe's IT* includes nor surefire's *Test patterns.

Each class gets its own JVM and therefore its own SparkSession, which
costs a little startup per class but lets the existing forks run the
groups concurrently.
